Changing order types to something more suitable

+++ VideoDatabase/Tables/VideoClips.sql
@@ -3,7 +3,7 @@
 	[VideoClipId] INT IDENTITY NOT NULL PRIMARY KEY,
 	[ClipId] INT NOT NULL,
 	[VideoId] INT NOT NULL,
-	[Order] INT NOT NULL,
+	[Order] SMALLINT NOT NULL,
 	[DateCreated] DATETIME2 NOT NULL,
 	CONSTRAINT FK_VideoClips_Clip FOREIGN KEY ([ClipId]) REFERENCES [dbo].[Clip] ([ClipId]),
 	CONSTRAINT FK_VideoClips_Video FOREIGN KEY ([VideoId]) REFERENCES [dbo].[Video] ([VideoId]),
